Aurel 9.9

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -11,17 +11,19 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->string('fullname', 50);
-            $table->string('password', 10);
-            $table->enum('kelas', ['B1', 'B2', 'B3']);
-            $table->text('catatan')->nullable();
-            $table->unsignedBigInteger('stiker_id')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->string('fullname', 50);
+                $table->string('password', 10);
+                $table->enum('kelas', ['B1', 'B2', 'B3']);
+                $table->text('catatan')->nullable();
+                $table->unsignedBigInteger('stiker_id')->nullable();
+                $table->timestamps();
 
-            $table->foreign('stiker_id')->references('id')->on('stiker')->onDelete('cascade');
-        });
+                $table->foreign('stiker_id')->references('id')->on('stiker')->onDelete('cascade');
+            });
+        }
     }
 
     /**
